// SofTec_TMS/NWC.Web/src/app/TMS-Module/Models/vehicle-data.model.ts
export class VehicleData {

    VehicleId: string;
    Code: string;
    PlateNo: string;
    Capacity: number;
    PermissionExpiryDate: Date;
    chassisNo: string;
    licenseExpiryDate: Date;
    CityId: string;
    CityName: string;
    AreaId: string;
    AreaName: string;
    StationId: string;
    StationName: string;
    DriverId: string;
    DriverName: string;
    DriverMobile: string;
    DriverCode: string;
    DriverPersonalId: string;
    StatusId: number;
    StatusName: string;
    StatusNameAr: string;
    StatusColorCode: string;
    ManufacturerId: string;
    ManufacturerName: string;
    ManufacturerNameAr: string;
    BrandId: string;
    BrandName: string;
    BrandNameAr: string;
    ProductionYearId: string;
    ProductionYearName: string;
    TypeId: string;
    TypeName: string;
    TypeNameAr: string;
    ServiceTypeID: number;
    ServiceTypeEN: string;
    ServiceTypeAR: string;
    ClassesAr: string;
    ClassesEn: string;

}